Reimplementing HbAVSS light
HbAVSS light has been reimplemented for the purpose of offline phase.
This also does not assume that the dealer is outside of the parties that are
participating which is the case in the offline phase. The client mode is
also supported where the dealer is not one of the parties
participating in the MPC as was the case with the previous version.