Implent type_check decorators for dynamically checking types
This introduces 3 new decorators-- there's the main type_check, which
largely works as expected for instance methods, as well as
static_type_check, which works for static methods / functions, and
class_type_check, which works for classmethods. This also implements
some type checking throughout MPC code to demonstrate the usage.