Refactored type checking into a class
This commit refactors type checking to use class-based decorators, which
are easier to read and maintain. Furthermore, this new approach means
that type checking is called via assertion, thus, when we run with
python -O, type checking is disabled.

This also ports the typechecking into a single annotation which allows
arithmetic typechecking, and forcing typechecking even when run in debug
mode.

Finally, this refactors typechecking to operate on method annotations
instead of decorator arguments.